more folder cleanup, rename hooks, add color cycle to easter eggs, tweak scoreboard coloring
