+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "@strive/hackathon-website",
+  "tagline": "The world is your canvas.",
   "private": true,
   "version": "0.0.0",
   "scripts": {
