+++ src/lib/Layout.tsx
@@ -1,19 +1,19 @@
-import { useMemo } from 'react';
+import { useContext, useMemo } from 'react';
 import { MoonIcon, SunIcon } from '@heroicons/react/solid';
-import { useDarkMode } from './use-dark';
-import DropdownMenu from './DropdownMenu';
+import { useDarkMode } from './useDark';
+import ThumbMenu from './ThumbMenu';
 import logoUrl from '../assets/logo.png';
 import lightLogo from '../assets/logo_light.png';
+import { BonusContext } from '../App';
 
-export function AppHeader() {
+export function AppHeader({className = ''}: {className:string}) {
   const { darkMode, toggle } = useDarkMode();
-
   const Logo = () => useMemo(() => (
     <div><img src={darkMode ? lightLogo : logoUrl} className='h-8 w-auto translate-y-1 opacity-90'/></div>
   ), [darkMode, toggle])
 
   return (
-    <header className="App-header">
+    <header className={className}>
       <div className="container flex">
         <a href="https://striveconsulting.com/" target="_blank" className="flex-1 pb-1.5">
           <div className='flex items-center gap-2 hover:-translate-y-1 transition-all duration-250'>
@@ -28,11 +28,24 @@ export function AppHeader() {
           <button onClick={() => toggle(!darkMode)} type="button" className="flex-1 ui !bg-transparent">
             {darkMode ? <MoonIcon className="w-5 h-5 text-white" /> : <SunIcon className="w-5 h-5 text-white" />}
           </button>
-          <DropdownMenu />
+          <ThumbMenu />
         </nav>
       </div>
     </header>
   );
 }
 
-export default { AppHeader };
+
+export function Layout({children}:React.PropsWithChildren<{}>) {
+  const {bonus} = useContext(BonusContext);
+  return (
+    <div className={`App-backdrop`}>
+      <AppHeader className='App-header' />
+      <main className={`App-container ${bonus ? 'electric-dream' : ''}`}>
+        {children}
+      </main>
+    </div>
+  )
+}
+
+export default { Layout, AppHeader };
