+++ src/styles/index.css
@@ -1,3 +1,8 @@
+/**
+ * This file is loaded immediately and will handle styles until our app is ready. Do not style
+ * the app itself from this file, use `tailwind.css` or a relevant component.css instead.
+ **/
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
