+++ src/styles/minesweeper.css
@@ -9,23 +9,26 @@
     @apply bg-red-300;
   }
   button.ms-item {
-    @apply w-6 h-6 sm:w-8 sm:h-8 lg:w-10 lg:h-10 sm:text-lg lg:text-xl rounded font-light;
-    @apply shadow bg-white/25 hover:scale-110 hover:shadow-lg;
-    @apply hover:bg-white/25 transition-all duration-500 ease-in-out;
+    @apply w-6 h-6 sm:w-8 sm:h-8 lg:w-10 lg:h-10 ;
+    @apply sm:text-lg lg:text-xl rounded font-light;
+    @apply shadow bg-white/25;
+    @apply transition-all duration-500 ease-in-out;
 
     &.close {
-      @apply text-transparent;
+      @apply text-transparent hover:bg-white/25;
+      @apply hover:scale-110 hover:shadow-lg;
     }
 
     &.open {
+      @apply cursor-default;
       @apply text-white;
       @apply bg-white/5;
       &.count-0 { @apply text-white/10 }
-      &.count-1 { @apply text-white }
-      &.count-2 { @apply text-orange-100 bg-orange-200/10 }
-      &.count-3 { @apply text-orange-200 bg-orange-300/10 }
-      &.count-4 { @apply text-orange-300 bg-orange-400/10 }
-      &.count-5 { @apply text-orange-400 bg-orange-500/10 }
+      &.count-1 { @apply text-white/60 bg-orange-300/10 }
+      &.count-2 { @apply text-white/75 bg-orange-300/20 }
+      &.count-3 { @apply text-white bg-orange-400/20 }
+      &.count-4 { @apply text-orange-200 bg-orange-500/20 }
+      &.count-5 { @apply text-orange-300 bg-orange-600/20 }
     }
 
   }
